+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+1.0.0-beta10
+------------
+- Fixed a bug in logging where the clj and cljs API didn't match
+- Added `defsc` macro to core namespace and documented in devguide section M05
+- Added a cheat sheet to docs
+- Added mutation helper comments to routing section of devguide
+- Enhanced the `env` available in post mutations and fallbacks
+- Added some content to the reference guide
+- Added auto-retries on dynamic route loading
+
 1.0.0-beta9
 -----------
 - Fixed bugs in gettext defaults for i18n extract/deploy.